mri_classifier_package_v2_release/requirements.txt
torch>=2.0
numpy>=1.24
nibabel>=5.0
matplotlib>=3.7
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
tomli>=2.0
deepbet>=1.0
shap>=0.42
